refactor: move thumbnail discovery out of the SDK
Review was right that this does not belong in sdk. Choosing a value for
metadata.thumbnail is metadata construction, not part of the tunnelling
contract, and the SDK should receive a resolved LeaseMetadata and stay
agnostic about where the value came from.

Keeping it there cost more than it looked: ThumbnailFromTarget on the
public ExposeConfig, DiscoverThumbnail exported as a network-fetching API
that every SDK consumer inherits, and HTML fetching and parsing inside the
transport layer. sdk/ is now untouched by this branch -- zero lines.

It also reframes the CodeQL finding correctly. In sdk it read as "the SDK
issues requests to caller-supplied addresses"; here it is what it actually
is, an explicit CLI action against the target the operator just named on
the command line, which the tunnel proxies every byte to anyway.

The discovery now lives in cmd/portal-tunnel/thumbnail, shared by
`portal expose` and the agent, and both resolve the value before handing
metadata to sdk.Expose. The agent resolves only when a tunnel starts:
metadataFromTunnelConfig stays pure, so the metadata-update and snapshot
paths do not re-read the target every time.

Taking the reviewer's follow-on simplification too: only absolute http(s)
image URLs are accepted. Open Graph asks for one and metadata.thumbnail is
absolute by contract, so resolving a relative reference meant deriving a
lease hostname from a chosen relay -- coupling a card image to relay
selection, and leaving an arbitrary choice to document when several relays
are configured. All of that is gone: no publicHostname, no LeaseHostname,
no PortalRootHost, and the reviewer's open question about explicitRelayURLs
versus relaySetURLs no longer exists.

Two more from review while the code moved:

  - Redirects are no longer followed. The target's front page is the whole
    request; a redirect would send it somewhere the operator did not name,
    so it is returned as a response and rejected by the status check.

  - The flag and the three documentation sites said "og:image" while the
    implementation also falls back to twitter:image and icon links. They
    now state the order, and that only absolute URLs are taken.
